test: move tests to their dirs
